Use createGlobalStyle for global styles
Use createGlobalStyle for global styles
